Move avg_color and win_pct_color to components/util.rs
These are general color-helper utilities, not specific to the boxscore
module. Expose util as pub mod so the helpers are accessible from ui/.
