+++ src/components/boxscore.rs
@@ -1,4 +1,5 @@
 use crate::components::game::player::Player;
+use crate::components::util::avg_color;
 use crate::state::app_state::HomeOrAway;
 use mlbt_api::boxscore::{LabelValue, Player as ApiPlayer, Team};
 use mlbt_api::live::LiveResponse;
@@ -9,32 +10,6 @@ use tui::text::Span;
 use tui::widgets::Cell;
 
 const SECONDARY_COLOR: Color = Color::DarkGray;
-
-pub(crate) fn win_pct_color(pct: &str) -> Option<Color> {
-    pct.parse::<f64>().ok().map(|v| {
-        if v == 0.0 {
-            Color::DarkGray
-        } else if v >= 0.500 {
-            Color::Green
-        } else {
-            Color::Red
-        }
-    })
-}
-
-pub(crate) fn avg_color(avg: &str) -> Option<Color> {
-    avg.parse::<f64>().ok().map(|v| {
-        if v == 0.0 {
-            Color::DarkGray
-        } else if v >= 0.300 {
-            Color::Green
-        } else if v < 0.100 {
-            Color::Red
-        } else {
-            Color::White
-        }
-    })
-}
 
 #[derive(Default)]
 pub struct Boxscore {
