+++ src/components/mod.rs
@@ -10,4 +10,4 @@ pub mod schedule;
 pub mod standings;
 pub mod stats;
 pub mod team_page;
-mod util;
+pub mod util;
